pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "production-rag-pipeline"
version = "1.0.0"
description = "Production-oriented RAG search pipeline with Bing, DuckDuckGo, extraction, chunking, and reranking."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.8"
authors = [
  { name = "KazKozDev" }
]
dependencies = [
  "beautifulsoup4>=4.12.0",
  "curl-cffi>=0.7.0",
  "PyYAML>=6.0.0"
]

[project.optional-dependencies]
extraction = [
  "trafilatura>=1.6.0"
]
semantic = [
  "numpy>=1.24.0",
  "scikit-learn>=1.3.0",
  "sentence-transformers>=2.2.0"
]
full = [
  "trafilatura>=1.6.0",
  "numpy>=1.24.0",
  "scikit-learn>=1.3.0",
  "sentence-transformers>=2.2.0"
]

[project.urls]
Homepage = "https://github.com/KazKozDev/production_rag_pipeline"

[project.scripts]
production-rag-pipeline = "production_rag_pipeline.cli:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
